requirements.txt
opencv-python==4.8.0.76
numpy==1.24.4
streamlit==1.25.0
Pillow==10.0.0
matplotlib==3.8.0
